requirements.txt
# Core Dependencies
streamlit==1.29.0
python-dotenv==1.0.0
Pillow>=10.2.0
requests==2.31.0

# API Clients
openai>=1.12.0
groq>=0.4.0
google-generativeai>=0.3.0

# Image Processing
opencv-python==4.8.1.78
numpy>=1.26.0

# Async Support
aiohttp==3.9.1